requirements.txt
# FastAPI backend
fastapi==0.104.1
pydantic==2.5.0
uvicorn==0.24.0

# Streamlit UI
streamlit
anthropic
python-dotenv
